+++ packages/version/index.js
@@ -1,5 +1,5 @@
 module.exports = {
   "decorator": "3.20.0",
   "core": "3.20.0",
-  "timestamp": 1737127217644
+  "timestamp": 1738332637497
 };